Fix re-entering a webform (again).
Trigger the code in webform_client_form() that is used to restart a
draft submission, because that’s exactly what we want to do.

It’s all about passing the right arguments to webform_client_form().
Since these change once in a while we need to narrow down the supported
webform versions.